Debugging on main loop
Continuing debugging for implementation of midi, math synth, playlist and channels, as well as audio output. Currently solved issue regarding poor range loop usage, resulting in redundant variable modification not passing to runtime variables

+++ SAMPLETOY_ESP32_Arduino/src/SampletoyPlaylist.c
@@ -99,11 +99,11 @@ void miditrackAddMidiEvent(miditrack* target, midinote event){
 //playlist methods
 int recalculateSamplesPerSubdivision(uint8_t new_bpm){
     //function is kind of inefficient on memory, but i'm assuming its not happening in runtime so i'ma write it more readable
-    uint16_t subdivisions_per_minute = new_bpm * SUBDIVISIONS_PER_BEAT;
+    int subdivisions_per_minute = new_bpm * SUBDIVISIONS_PER_BEAT;
     float sub_per_second = subdivisions_per_minute/60;
     float time_for_one_sub = 1/sub_per_second;
     //typecast time for one subdivision into int, after turning it into x number of samples that fit into it
-    int samples_per_sub = (int) time_for_one_sub / SAMPLE_DURATION;
+    int samples_per_sub = (int) (time_for_one_sub / SAMPLE_DURATION);
     return samples_per_sub;
 }
